requirements.txt
Django==1.8.2
django-crispy-forms==1.4.0
django-discover-runner==1.0
django-role-permissions==0.4.2
gunicorn==19.3.0
psycopg2==2.6
